server/requirements.txt
watchfiles
sqlalchemy
pydantic
uvicorn
bcrypt
pyjwt